+++ packages/trader/src/AppV2/Components/AccumulatorStats/accumulator-stats-row.tsx
