+++ packages/trader/src/AppV2/Components/AccumulatorStats/index.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
 import './stats.scss';
-import Stats from './stats';
+import Stats from './accumulator-stats';
 
 export default Stats;
