+++ packages/trader/src/AppV2/Components/AccumulatorStats/stats.scss
